Smoother
Included MovingAverage.h to Smoother.cpp
