+++ src/Smoother.h
@@ -1,0 +1,47 @@
+/*
+ * Smoother.h
+ *
+ * (c) 2015 Sofian Audry        :: info(@)sofianaudry(.)com
+ * (c) 2015 Thomas O Fredericks :: tof(@)t-o-f(.)info
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef SMOOTHER_H_
+#define SMOOTHER_H_
+
+#include "PqCore.h"
+
+/// Simple moving average transform filter.
+class Smoother : public PqPutter, public MovingAverage {
+public:
+  /**
+   * Constructor.
+   * @param factor a parameter in [0, 1] representing the importance of new values as opposed to old values (ie. lower smoothing factor means *more* smoothing)
+   */
+  Smoother(float factor=0.1f);
+  virtual ~Smoother() {}
+
+  /**
+   * Pushes value into the unit.
+   * @param value the value sent to the unit
+   * @return the new value of the unit
+   */
+  virtual float put(float value);
+
+  /// Returns smoothed value.
+  virtual float get() { return MovingAverage::get(); }
+};
+
+#endif
